refactor: extracts bootstrap side effect into server.ts file
