+++ apps/tx-signer/package.json
@@ -12,7 +12,7 @@
     "dev-nodc": "npm run start",
     "format": "prettier --write ./*.{ts,js,json} **/*.{ts,js,json}",
     "lint": "eslint .",
-    "prod": "node --require ./dist/instrumentation dist/server.js",
+    "prod": "node --require ./dist/instrumentation.js ./dist/server.js",
     "start": "tsup --watch",
     "start:dev": "tsup src/server.ts --watch",
     "test": "vitest run --project unit --project functional",
@@ -25,6 +25,7 @@
     "@akashnetwork/chain-sdk": "1.0.0-alpha.25",
     "@akashnetwork/env-loader": "*",
     "@akashnetwork/http-sdk": "*",
+    "@akashnetwork/instrumentation": "*",
     "@akashnetwork/logging": "*",
     "@cosmjs/amino": "~0.36.1",
     "@cosmjs/crypto": "~0.36.1",
@@ -36,7 +37,6 @@
     "@hono/otel": "~0.4.0",
     "@hono/zod-openapi": "0.18.4",
     "@opentelemetry/api": "^1.9.0",
-    "@akashnetwork/instrumentation": "*",
     "async-sema": "^3.1.1",
     "cockatiel": "^3.2.1",
     "dataloader": "^2.2.2",
@@ -59,10 +59,10 @@
     "@types/node": "^22.13.11",
     "@types/uuid": "^8.3.1",
     "@typescript-eslint/eslint-plugin": "^7.12.0",
+    "@vitest/coverage-v8": "^4.0.0",
     "eslint": "^8.57.0",
     "eslint-config-next": "^14.2.3",
     "eslint-plugin-simple-import-sort": "^12.1.0",
-    "@vitest/coverage-v8": "^4.0.0",
     "prettier": "^3.3.0",
     "prettier-plugin-tailwindcss": "^0.6.1",
     "tsup": "^8.5.1",
